Bring driver and notifier into an SNSDriver
